+++ goblin-together/src/main/java/com/doggybear/Main.java
@@ -5,18 +5,24 @@ import com.almasb.fxgl.app.GameSettings;
 import com.almasb.fxgl.app.scene.FXGLMenu;
 import com.almasb.fxgl.app.scene.SceneFactory;
 import com.almasb.fxgl.entity.Entity;
+import com.almasb.fxgl.entity.SpawnData;
+import com.almasb.fxgl.input.UserAction;
+import com.almasb.fxgl.physics.CollisionHandler;
 import com.almasb.fxgl.physics.PhysicsComponent;
+import com.almasb.fxgl.physics.PhysicsWorld;
 import com.doggybear.component.Goblin;
 import com.doggybear.component.Timer;
-import com.doggybear.controller.GameController;
-import com.doggybear.controller.InputController;
-import com.doggybear.controller.PhysicsController;
-import com.doggybear.controller.UIController;
+import com.doggybear.type.EntityType;
+
+import com.doggybear.factory.FactoryManager;
+
+import com.doggybear.levels.Level;
+import com.doggybear.levels.LevelManager;
 import com.doggybear.menu.MainMenu;
-import com.doggybear.network.NetworkGameManager;
 
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
@@ -27,15 +33,6 @@ import javafx.scene.text.Text;
 import static com.almasb.fxgl.dsl.FXGL.*;
 
 public class Main extends GameApplication {
-    
-    // 控制器
-    private GameController gameController;
-    private InputController inputController;
-    private PhysicsController physicsController;
-    private UIController uiController;
-    
-    // 遊戲更新相關
-    private double timePassed = 0;
     
     @Override
     protected void initSettings(GameSettings settings) {
@@ -52,101 +49,238 @@ public class Main extends GameApplication {
         });
     }
 
+    private Entity goblin;
+    private Entity goblin2; // 新增第二個哥布林
+    private Timer timer;
+
+    private Entity lava;
+    private double lavaHeight = 100; // 初始岩漿高度
+    private double lavaRiseSpeed = 5; // 岩漿上升速度
+    private double timePassed = 0;
+    private double lavaY = 1000; // 岩漿底部的固定Y坐標
+
+    private boolean isGameOver = false;
+    private int WORLD_HEIGHT = 10000; // 遊戲世界的總高度
+
+    private Level level;
+
     @Override
     protected void initGame() {
-        System.out.println("=== Main.initGame() 開始 ===");
+        isGameOver = false;
+        timePassed = 0;
         
-        // 初始化遊戲控制器
-        gameController = new GameController();
-        gameController.initGame();
+        FactoryManager.addAllFactories(getGameWorld());
         
-        // 初始化物理控制器
-        physicsController = new PhysicsController(
-            gameController.getNetworkGameManager(),
-            this::showGameOver  // 遊戲結束回調
-        );
-        physicsController.setGravity(0, 1500);
+        getGameScene().setBackgroundColor(Color.LIGHTBLUE);
         
-        // 初始化UI控制器
-        uiController = new UIController(
-            gameController.getNetworkGameManager(),
-            gameController.getLevel()
-        );
+        level = LevelManager.createLevel();
         
-        // 在這裡直接初始化輸入控制器
-        System.out.println("在 initGame() 中初始化輸入...");
-        Entity goblin = gameController.getGoblin();
-        Entity goblin2 = gameController.getGoblin2();
+        // 生成第一個哥布林
+        goblin = spawn("goblin", level.getGoblinStartX(), level.getGoblinStartY());
         
-        if (goblin != null && goblin2 != null) {
-            inputController = new InputController(
-                goblin,
-                goblin2,
-                gameController.getNetworkGameManager()
-            );
-            
-            System.out.println("調用 InputController.initInput()...");
-            inputController.initInput();
-        } else {
-            System.err.println("錯誤：無法在 initGame() 中初始化輸入，實體為 null");
-        }
+        // 生成第二個哥布林
+        goblin2 = spawn("goblin2", level.getGoblin2StartX(), level.getGoblin2StartY());
         
-        System.out.println("=== Main.initGame() 完成 ===");
+        timer = new Timer();
+        goblin.addComponent(timer);
+        
+        lavaHeight = level.getInitialLavaHeight();
+        lavaRiseSpeed = level.getLavaRiseSpeed();
+        
+        lava = spawn("lava", new SpawnData(0, lavaY - lavaHeight)
+              .put("width", (int)getAppWidth())
+              .put("height", (int)lavaHeight));
+        
+        getPhysicsWorld().setGravity(0, 1000);
+        
+        // 修改視角，確保兩個玩家都在畫面中
+        getGameScene().getViewport().setBounds(0, -WORLD_HEIGHT, getAppWidth(), WORLD_HEIGHT + getAppHeight());
+        
+        // 動態調整視角以確保兩個玩家都在畫面中
+        updateViewport();
+    }
+
+    // 更新視角
+    private void updateViewport() {
+        if (goblin == null || goblin2 == null) return;
+        
+        // 計算兩個哥布林的中心點作為視角中心
+        double centerX = (goblin.getX() + goblin2.getX()) / 2 + 25; // 加上一半的寬度(50/2)
+        double centerY = (goblin.getY() + goblin2.getY()) / 2 + 25; // 加上一半的高度(50/2)
+        
+        // 計算目標視角位置（讓中心點位於畫面中央）
+        double targetViewX = centerX - getAppWidth() / 2;
+        double targetViewY = centerY - getAppHeight() / 2;
+        
+        targetViewX = 0; // 固定X軸位置
+        
+        double minViewY = -WORLD_HEIGHT; // 上邊界
+        double maxViewY = 0; // 下邊界
+        
+        targetViewY = Math.max(minViewY, Math.min(targetViewY, maxViewY));
+        
+        getGameScene().getViewport().setX(targetViewX);
+        getGameScene().getViewport().setY(targetViewY);
     }
 
     @Override
     protected void initInput() {
-        // 確保遊戲控制器已經初始化且實體已創建
-        if (gameController == null) {
-            System.err.println("GameController 未初始化！");
-            return;
-        }
-        
-        Entity goblin = gameController.getGoblin();
-        Entity goblin2 = gameController.getGoblin2();
-        
-        if (goblin == null || goblin2 == null) {
-            System.err.println("Goblin 實體未創建！");
-            return;
-        }
-        
-        // 現在創建輸入控制器
-        inputController = new InputController(
-            goblin,
-            goblin2,
-            gameController.getNetworkGameManager()
-        );
-        
-        // 初始化輸入
-        inputController.initInput();
-        
-        System.out.println("輸入控制器初始化完成");
+        // 第一个玩家控制 - WASD 和空格跳躍
+        getInput().addAction(new UserAction("向右移動") {
+            @Override
+            protected void onAction() {
+                goblin.getComponent(Goblin.class).moveRight();
+            }
+
+            @Override
+            protected void onActionEnd() {
+                goblin.getComponent(Goblin.class).stop();
+            }
+        }, KeyCode.D);
+
+        getInput().addAction(new UserAction("向左移動") {
+            @Override
+            protected void onAction() {
+                goblin.getComponent(Goblin.class).moveLeft();
+            }
+
+            @Override
+            protected void onActionEnd() {
+                goblin.getComponent(Goblin.class).stop();
+            }
+        }, KeyCode.A);
+
+        getInput().addAction(new UserAction("跳躍") {
+            @Override
+            protected void onActionBegin() {
+                goblin.getComponent(Goblin.class).jump();
+            }
+        }, KeyCode.SPACE);
+
+        // 第二个玩家控制 - 方向鍵和Enter跳躍
+        getInput().addAction(new UserAction("玩家2向右移動") {
+            @Override
+            protected void onAction() {
+                goblin2.getComponent(Goblin.class).moveRight();
+            }
+
+            @Override
+            protected void onActionEnd() {
+                goblin2.getComponent(Goblin.class).stop();
+            }
+        }, KeyCode.RIGHT);
+
+        getInput().addAction(new UserAction("玩家2向左移動") {
+            @Override
+            protected void onAction() {
+                goblin2.getComponent(Goblin.class).moveLeft();
+            }
+
+            @Override
+            protected void onActionEnd() {
+                goblin2.getComponent(Goblin.class).stop();
+            }
+        }, KeyCode.LEFT);
+
+        getInput().addAction(new UserAction("玩家2跳躍") {
+            @Override
+            protected void onActionBegin() {
+                goblin2.getComponent(Goblin.class).jump();
+            }
+        }, KeyCode.ENTER);
     }
     
     @Override
     protected void initPhysics() {
-        physicsController.initPhysics();
-    }
-    
-    @Override
-    protected void initUI() {
-        // 確保控制器已經初始化
-        if (uiController != null) {
-            uiController.initUI();
-        }
+        PhysicsWorld physicsWorld = getPhysicsWorld();
+        
+        // 第一个玩家的碰撞处理
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN, EntityType.PLATFORM) {
+            @Override
+            protected void onCollisionBegin(Entity goblin, Entity platform) {
+                goblin.getComponent(Goblin.class).onGroundCollision();
+            }
+        });
+
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN, EntityType.LAVA) {
+            @Override
+            protected void onCollisionBegin(Entity goblin, Entity lava) {
+                showGameOver();
+            }
+        });
+        
+        // 第二个玩家的碰撞处理
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN2, EntityType.PLATFORM) {
+            @Override
+            protected void onCollisionBegin(Entity goblin2, Entity platform) {
+                goblin2.getComponent(Goblin.class).onGroundCollision();
+            }
+        });
+
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN2, EntityType.LAVA) {
+            @Override
+            protected void onCollisionBegin(Entity goblin2, Entity lava) {
+                showGameOver();
+            }
+        });
+        
+        // 刺的碰撞处理
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN, EntityType.SPIKE) {
+            @Override
+            protected void onCollisionBegin(Entity goblin, Entity spike) {
+                showGameOver();
+            }
+        });
+        
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN2, EntityType.SPIKE) {
+            @Override
+            protected void onCollisionBegin(Entity goblin2, Entity spike) {
+                showGameOver();
+            }
+        });
+
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN, EntityType.ARROW) {
+            @Override
+            protected void onCollisionBegin(Entity goblin, Entity arrow) {
+                // 移除弓箭
+                arrow.removeFromWorld();
+                // 触发游戏结束
+                showGameOver();
+            }
+        });
+        
+        // 第二个玩家被弓箭击中
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.GOBLIN2, EntityType.ARROW) {
+            @Override
+            protected void onCollisionBegin(Entity goblin2, Entity arrow) {
+                // 移除弓箭
+                arrow.removeFromWorld();
+                // 触发游戏结束
+                showGameOver();
+            }
+        });
+        
+        // 弓箭击中平台时消失
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.ARROW, EntityType.PLATFORM) {
+            @Override
+            protected void onCollisionBegin(Entity arrow, Entity platform) {
+                arrow.removeFromWorld();
+            }
+        });
+        
+        // 弓箭击中刺时消失（可选）
+        physicsWorld.addCollisionHandler(new CollisionHandler(EntityType.ARROW, EntityType.SPIKE) {
+            @Override
+            protected void onCollisionBegin(Entity arrow, Entity spike) {
+                arrow.removeFromWorld();
+            }
+        });
     }
     
     @Override
     protected void onUpdate(double tpf) {
-        // 確保控制器已初始化
-        if (gameController == null) return;
-        
-        Entity goblin = gameController.getGoblin();
-        Entity goblin2 = gameController.getGoblin2();
-        com.doggybear.component.Timer timer = gameController.getTimer();
-        NetworkGameManager networkGameManager = gameController.getNetworkGameManager();
-        
-        if (gameController.isGameOver() || gameController.isClientGameOver()) {
+        // 如果遊戲已經結束 不再更新岩漿 Goblin 禁止移動
+        if (isGameOver) {
             if (goblin != null && goblin.isActive()) {
                 goblin.getComponent(Goblin.class).stop();
                 goblin.getComponent(PhysicsComponent.class).setVelocityY(0);
@@ -158,84 +292,41 @@ public class Main extends GameApplication {
             return;
         }
         
-        if (networkGameManager.isNetworkGame()) {
-            // 獲取當前存活時間
-            int currentSurvivalTime = timer != null ? timer.getElapsedSeconds() : 0;
-            
-            // 更新網路遊戲（包含插值和狀態同步）
-            networkGameManager.updateNetworkGame(tpf, goblin, goblin2, gameController.getLavaHeight(), currentSurvivalTime);
-            
-            // 只有主機端執行遊戲邏輯
-            if (networkGameManager.getNetworkManager().isHost()) {
-                gameController.updateGameLogic(tpf);
-                
-                // 主機端檢查遊戲結束條件
-                if (gameController.checkGameOverConditions()) {
-                    showGameOver();
-                }
-            } else {
-                // 客戶端只更新岩漿視覺效果（根據主機端同步的數據）
-                gameController.updateClientLava();
-            }
-        } else {
-            // 本地遊戲，正常執行所有邏輯
-            gameController.updateGameLogic(tpf);
-            
-            if (gameController.checkGameOverConditions()) {
-                showGameOver();
-            }
-        }
-        
-        // 更新視窗
-        gameController.updateViewport();
-        
-        // 更新除錯資訊（可選）
-        updateDebugInfo(tpf);
-    }
-    
-    private void updateDebugInfo(double tpf) {
-        // 確保 UI 控制器已初始化
-        if (uiController == null || gameController == null) return;
-        
-        // 每秒更新一次除錯資訊
+        // 更新經過的時間
         timePassed += tpf;
-        if (timePassed >= 1.0) {
-            double fps = 1.0 / tpf;
-            int entityCount = getGameWorld().getEntities().size();
-            String networkInfo = gameController.getNetworkGameManager().isNetworkGame() ? 
-                "網路: " + (gameController.getNetworkGameManager().getNetworkManager().isHost() ? "主機" : "客戶端") :
-                "本地遊戲";
-                
-            uiController.showDebugInfo(fps, entityCount, networkInfo);
+        
+        // 每0.5秒更新岩漿高度，讓它變高
+        if (timePassed > 0.5) {
+            lavaHeight += lavaRiseSpeed;
+            
+            // 重新設置岩漿的Y坐標和高度
+            lava.removeFromWorld();
+            lava = spawn("lava", new SpawnData(0, lavaY - lavaHeight)
+                .put("width", (int)getAppWidth())
+                .put("height", (int)lavaHeight));
+            
             timePassed = 0;
         }
+        
+        // 檢查兩位玩家是否都掉入岩漿
+        if ((goblin.getY() + goblin.getHeight() > lavaY - lavaHeight) || 
+            (goblin2.getY() + goblin2.getHeight() > lavaY - lavaHeight)) {
+            showGameOver();
+        }
+        
+        // 更新視角位置，使兩個玩家都保持在畫面中
+        updateViewport();
     }
 
     private void showGameOver() {
-        if (gameController.isGameOver() || gameController.isClientGameOver()) return;
+        if (isGameOver) return;
         
-        NetworkGameManager networkGameManager = gameController.getNetworkGameManager();
-        com.doggybear.component.Timer timer = gameController.getTimer();
-        
-        if (networkGameManager.isNetworkGame() && !networkGameManager.getNetworkManager().isHost()) {
-            gameController.setClientGameOver(true);
-        } else {
-            gameController.setGameOver(true);
-        }
-        
-        if (networkGameManager.isNetworkGame() && networkGameManager.getNetworkManager().isHost()) {
-            networkGameManager.sendGameOver();
-        }
+        isGameOver = true;
         
         timer.stop();
         
-        // 獲取最終存活時間
-        int finalSurvivalTime;
-        if (networkGameManager.isNetworkGame() && !networkGameManager.getNetworkManager().isHost()) {
-            finalSurvivalTime = gameController.getClientSurvivalTime(); // 客戶端使用主機端的時間
-        } else {
-            finalSurvivalTime = timer.getElapsedSeconds();
-        }
+        // 記錄最終存活時間
+        int finalSurvivalTime = timer.getElapsedSeconds();
         
         Rectangle overlay = new Rectangle(getAppWidth(), getAppHeight(), Color.color(0, 0, 0, 0.7));
         
@@ -249,13 +340,7 @@ public class Main extends GameApplication {
         gameOverText.setFont(Font.font(40));
         gameOverText.setFill(Color.WHITE);
         
-        String gameTypeText = networkGameManager.isNetworkGame() ? "網路合作" : "本地合作";
-        String roleText = "";
-        if (networkGameManager.isNetworkGame()) {
-            roleText = networkGameManager.getNetworkManager().isHost() ? " (主機)" : " (客戶端)";
-        }
-        
-        Text survivalTimeText = new Text(gameTypeText + roleText + "\n存活了 " + finalSurvivalTime + " 秒");
+        Text survivalTimeText = new Text("您存活了 " + finalSurvivalTime + " 秒");
         survivalTimeText.setFont(Font.font(20));
         survivalTimeText.setFill(Color.WHITE);
         
@@ -263,16 +348,14 @@ public class Main extends GameApplication {
         restartBtn.setPrefWidth(150);
         restartBtn.setPrefHeight(40);
         restartBtn.setOnAction(e -> {
-            networkGameManager.stopNetworkGame();
-            restartGame();
+            getGameController().startNewGame();
         });
         
         Button menuBtn = new Button("回到主頁面");
         menuBtn.setPrefWidth(150);
         menuBtn.setPrefHeight(40);
         menuBtn.setOnAction(e -> {
-            networkGameManager.stopNetworkGame();
-            getGameController().gotoMainMenu(); // 使用 FXGL 的 GameController
+            getGameController().gotoMainMenu();
         });
         
         VBox modalContent = new VBox(20);
@@ -285,34 +368,23 @@ public class Main extends GameApplication {
         
         getGameScene().addUINodes(overlay, modalPane);
     }
-    
-    /**
-     * 重新開始遊戲
-     */
-    private void restartGame() {
-        // 清理現有狀態
-        if (inputController != null) {
-            inputController.clearInput();
-        }
-        if (uiController != null) {
-            uiController.clearUI();
-        }
-        if (physicsController != null) {
-            physicsController.resetPhysics();
-        }
+
+    @Override
+    protected void initUI() {
+        Text title = new Text("Goblin Together");
+        title.setTranslateX(10);
+        title.setTranslateY(30);
         
-        // 重新開始遊戲 - 使用 FXGL 的 GameController
-        getGameController().startNewGame();
+        Text helpText = new Text("玩家1: A/D 移動，空白鍵跳躍 | 玩家2: 方向鍵移動，Enter跳躍");
+        helpText.setTranslateX(10);
+        helpText.setTranslateY(60);
+        
+        Text levelText = new Text("關卡: " + (level != null ? level.getName() : ""));
+        levelText.setTranslateX(10);
+        levelText.setTranslateY(90);
+        
+        getGameScene().addUINodes(title, helpText, levelText);
     }
-    
-    /**
-     * 獲取遊戲控制器實例（用於外部訪問）
-     * 重新命名避免與 FXGL 的 GameController 衝突
-     */
-    public GameController getMyGameController() { return gameController; }
-    public InputController getInputController() { return inputController; }
-    public PhysicsController getPhysicsController() { return physicsController; }
-    public UIController getUIController() { return uiController; }
 
     public static void main(String[] args) {
         launch(args);
